# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Velox API server.",
        "title": "Velox API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/api-keys": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve a list of all API keys created by the current user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APIKeys"
                ],
                "summary": "List API Keys",
                "responses": {
                    "200": {
                        "description": "List of API keys",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.APIKey"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new API key with specific scopes and optional expiration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APIKeys"
                ],
                "summary": "Generate API Key",
                "parameters": [
                    {
                        "description": "API Key Generation Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.GenerateKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "API key created",
                        "schema": {
                            "$ref": "#/definitions/handler.GenerateKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Permanently revoke and delete an API key.",
                "tags": [
                    "APIKeys"
                ],
                "summary": "Delete API Key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key UUID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Key deleted"
                    },
                    "400": {
                        "description": "Missing ID",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Rename an existing API key.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "APIKeys"
                ],
                "summary": "Update API Key Name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key UUID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Update Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Key updated"
                    },
                    "400": {
                        "description": "Missing ID or invalid JSON",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/auth/api-keys/stats": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve performance metrics and usage peaks for a specific API key.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APIKeys"
                ],
                "summary": "Get API Key Stats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "API Key Statistics",
                        "schema": {
                            "$ref": "#/definitions/model.APIKeyStats"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Authenticate a user and return a JWT token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User Login",
                "parameters": [
                    {
                        "description": "Login Request",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "$ref": "#/definitions/handler.successResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "description": "Invalidate the session (stateless JWT logout).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User Logout",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "schema": {
                            "$ref": "#/definitions/handler.successResponse"
                        }
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "description": "Register a new user with name, email, and password.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User Signup",
                "parameters": [
                    {
                        "description": "Signup Request",
                        "name": "signup",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.signupRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created",
                        "schema": {
                            "$ref": "#/definitions/handler.successResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetch profile and activity data for the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get Dashboard Data",
                "responses": {
                    "200": {
                        "description": "Dashboard data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/status": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get the current status and results of a code submission.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Judge"
                ],
                "summary": "Check Submission Status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Submission ID",
                        "name": "submission_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Submission Response (omitted if pending)",
                        "schema": {
                            "$ref": "#/definitions/judge.SubmissionResponse"
                        }
                    },
                    "400": {
                        "description": "Missing submission_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/submit": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Submit source code for execution against test cases.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Judge"
                ],
                "summary": "Submit Code",
                "parameters": [
                    {
                        "description": "Submission Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/judge.SubmissionRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "submission_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid JSON or limits too high",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to queue submission",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.GenerateKeyRequest": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "handler.GenerateKeyResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "display_hint": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                }
            }
        },
        "handler.UpdateKeyRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "handler.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "handler.loginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "handler.signupRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "handler.successResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "judge.SubmissionRequest": {
            "type": "object",
            "properties": {
                "language": {
                    "description": "e.g., \"cpp\", \"python\"",
                    "type": "string"
                },
                "memory_limit_kb": {
                    "type": "integer"
                },
                "source_code": {
                    "type": "string"
                },
                "submission_id": {
                    "type": "string"
                },
                "test_cases": {
                    "description": "Up to 20 cases",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/judge.TestCase"
                    }
                },
                "time_limit_ms": {
                    "type": "integer"
                }
            }
        },
        "judge.SubmissionResponse": {
            "type": "object",
            "properties": {
                "compile_error": {
                    "type": "string"
                },
                "overall_state": {
                    "description": "\"Accepted\", \"Wrong Answer\", \"Compile Error\"",
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/judge.TestCaseResult"
                    }
                },
                "submission_id": {
                    "type": "string"
                }
            }
        },
        "judge.TestCase": {
            "type": "object",
            "properties": {
                "expected_output": {
                    "type": "string"
                },
                "input": {
                    "type": "string"
                },
                "test_case_id": {
                    "type": "integer"
                }
            }
        },
        "judge.TestCaseResult": {
            "type": "object",
            "properties": {
                "actual_output": {
                    "type": "string"
                },
                "expected_output": {
                    "type": "string"
                },
                "input": {
                    "type": "string"
                },
                "memory_kb": {
                    "type": "integer"
                },
                "status": {
                    "description": "\"Passed\", \"Wrong Answer\", \"Runtime Error\"",
                    "type": "string"
                },
                "stderr": {
                    "type": "string"
                },
                "test_case_id": {
                    "type": "integer"
                },
                "time_ms": {
                    "type": "integer"
                }
            }
        },
        "model.APIKey": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "display_hint": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_used_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "model.APIKeyStats": {
            "type": "object",
            "properties": {
                "api_key_id": {
                    "type": "string"
                },
                "error_counts": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "peak_rpd": {
                    "type": "integer"
                },
                "peak_rpm": {
                    "type": "integer"
                },
                "success_rate": {
                    "type": "number"
                },
                "total_requests": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and then your personal token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}